Split validated_legacy_signed_public_key
