+++ xmtp_mls/src/credential/legacy_create_identity_association.rs
@@ -1,107 +1,14 @@
 use prost::Message;
-use xmtp_cryptography::signature::RecoverableSignature;
 use xmtp_proto::xmtp::{
     message_contents::{
-        signature::Union, signed_private_key, unsigned_public_key,
+        signed_private_key,
         SignedPrivateKey as LegacySignedPrivateKeyProto,
-        SignedPublicKey as LegacySignedPublicKeyProto,
-        UnsignedPublicKey as LegacyUnsignedPublicKeyProto,
     },
     mls::message_contents::LegacyCreateIdentityAssociation as LegacyCreateIdentityAssociationProto,
 };
 use xmtp_v2::k256_helper;
 
-use crate::types::Address;
-
-use super::AssociationError;
-
-struct ValidatedLegacySignedPublicKey {
-    account_address: Address,
-    serialized_key_data: Vec<u8>,
-    wallet_signature: RecoverableSignature,
-    public_key_bytes: Vec<u8>,
-    created_ns: u64,
-}
-
-impl ValidatedLegacySignedPublicKey {
-    fn header_text() -> String {
-        let label = "Create Identity".to_string();
-        format!("XMTP : {}", label)
-    }
-
-    fn body_text(serialized_legacy_key: &[u8]) -> String {
-        hex::encode(serialized_legacy_key)
-    }
-
-    fn footer_text() -> String {
-        "For more info: https://xmtp.org/signatures/".to_string()
-    }
-
-    fn text(serialized_legacy_key: &[u8]) -> String {
-        format!(
-            "{}\n{}\n\n{}",
-            Self::header_text(),
-            Self::body_text(serialized_legacy_key),
-            Self::footer_text()
-        )
-        .to_string()
-    }
-
-    pub fn account_address(&self) -> Address {
-        self.account_address.clone()
-    }
-
-    pub fn key_bytes(&self) -> Vec<u8> {
-        self.public_key_bytes.clone()
-    }
-
-    pub fn created_ns(&self) -> u64 {
-        self.created_ns
-    }
-}
-
-impl TryFrom<LegacySignedPublicKeyProto> for ValidatedLegacySignedPublicKey {
-    type Error = AssociationError;
-
-    fn try_from(proto: LegacySignedPublicKeyProto) -> Result<Self, AssociationError> {
-        let serialized_key_data = proto.key_bytes;
-        let Union::WalletEcdsaCompact(wallet_ecdsa_compact) = proto
-            .signature
-            .ok_or(AssociationError::MalformedLegacyKey)?
-            .union
-            .ok_or(AssociationError::MalformedLegacyKey)?
-        else {
-            return Err(AssociationError::MalformedLegacyKey);
-        };
-        let mut wallet_signature = wallet_ecdsa_compact.bytes.clone();
-        wallet_signature.push(wallet_ecdsa_compact.recovery as u8); // TODO: normalize recovery ID if necessary
-        let wallet_signature = RecoverableSignature::Eip191Signature(wallet_signature);
-        let account_address =
-            wallet_signature.recover_address(&Self::text(&serialized_key_data))?;
-        // TODO verify this is a legitimate address
-
-        let legacy_unsigned_public_key_proto =
-            LegacyUnsignedPublicKeyProto::decode(serialized_key_data.as_slice())
-                .or(Err(AssociationError::MalformedAssociation))?;
-        let public_key_bytes = match legacy_unsigned_public_key_proto
-            .union
-            .ok_or(AssociationError::MalformedAssociation)?
-        {
-            unsigned_public_key::Union::Secp256k1Uncompressed(secp256k1_uncompressed) => {
-                secp256k1_uncompressed.bytes
-            }
-        };
-        let created_ns = legacy_unsigned_public_key_proto.created_ns;
-
-        Ok(Self {
-            account_address,
-            wallet_signature,
-            serialized_key_data,
-            public_key_bytes,
-            created_ns,
-        })
-    }
-}
+use super::{validated_legacy_signed_public_key::ValidatedLegacySignedPublicKey, AssociationError};
 
 /// An Association is link between a blockchain account and an xmtp installation for the purposes of
 /// authentication.
@@ -136,7 +43,6 @@ impl LegacyCreateIdentityAssociation {
             .union
             .ok_or(AssociationError::MalformedLegacyKey)?;
         let legacy_private_key = secp256k1.bytes;
-        // TODO: Sign installation key using legacy private key
         let (mut delegating_signature, recovery_id) = k256_helper::sign_sha256(
             &legacy_private_key,      // secret_key
             &installation_public_key, // message
@@ -150,7 +56,7 @@ impl LegacyCreateIdentityAssociation {
         Self::new_validated(
             installation_public_key,
             delegating_signature,
-            legacy_signed_public_key_proto.try_into()?,
+            legacy_signed_public_key_proto.try_into()?, // ValidatedLegacySignedPublicKey
         )
     }
 
@@ -169,7 +75,7 @@ impl LegacyCreateIdentityAssociation {
         Self::new_validated(
             expected_installation_public_key.to_vec(),
             delegating_signature,
-            legacy_signed_public_key_proto.try_into()?,
+            legacy_signed_public_key_proto.try_into()?, // ValidatedLegacySignedPublicKey
         )
     }
 
@@ -205,11 +111,11 @@ impl LegacyCreateIdentityAssociation {
 
 #[cfg(test)]
 pub mod tests {
-    use ethers::signers::{LocalWallet, Signer};
-    use xmtp_cryptography::{signature::h160addr_to_string, utils::rng};
-    use xmtp_proto::xmtp::mls::message_contents::GrantMessagingAccessAssociation as GrantMessagingAccessAssociationProto;
+    
+    
+    
 
-    use crate::credential::grant_messaging_access::GrantMessagingAccessAssociation;
+    
 
     #[tokio::test]
     async fn assoc_gen() {}
